Delete confirm alert

+++ src/main/webapp/WEB-INF/views/siteUpdate.jsp
@@ -40,7 +40,7 @@
 									<li>
 									<a href="site?no=${cat.no}">${cat.name} &nbsp</a> 
 									<a href="categoryUpdate?no=${cat.no}&name=${cat.name}"><span class="glyphicon glyphicon-pencil"></span>&nbsp</a> 
-									<a href="categoryDeleteAction?no=${cat.no}"><span class="glyphicon glyphicon-remove"></span></a>
+									<a href="categoryDeleteAction?no=${cat.no}" onclick="return confirm('삭제하시겠습니까?');"><span class="glyphicon glyphicon-remove"></span></a>
 									</li>
 								</c:forEach>
 							</ul>
